Refactor billing endpoints
Pulls billing data endpoint into standalone endpoint under an establishment rather than nesting under PILs, which didn't make a huge amount of sense.
