+++ lib/routers/establishment/index.js
@@ -104,5 +104,6 @@ router.use('/:establishment/project(s)?', require('./projects'));
 router.use('/:establishment/invite-user', require('./invite-user'));
 router.use('/:establishment/invitations', require('./invitations'));
 router.use('/:establishment/pil(s)?', require('../profile/pil'));
+router.use('/:establishment/billing', require('./billing'));
 
 module.exports = router;
