+++ lib/routers/profile/pil.js
@@ -1,4 +1,4 @@
-const { get, pick } = require('lodash');
+const { get } = require('lodash');
 const moment = require('moment');
 const { NotFoundError, BadRequestError } = require('../../errors');
 const { fetchOpenTasks, permissions, validateSchema, whitelist, updateDataAndStatus } = require('../../middleware');
@@ -86,58 +86,6 @@ const attachEstablishmentDetails = (req, res, next) => {
     .then(() => next())
     .catch(next);
 };
-
-const clean = (id, start, end) => pil => {
-  // remove sensitive information from profile
-  pil.profile.establishments = pil.profile.establishments.filter(e => e.id === id);
-  pil.profile = pick(pil.profile, 'id', 'firstName', 'lastName', 'establishments');
-
-  pil.startDate = pil.pilTransfers
-    .filter(t => t.toEstablishmentId === id)
-    .reduce((date, t) => {
-      return (t.createdAt > date && t.createdAt < end) ? t.createdAt : date;
-    }, pil.issueDate);
-
-  pil.endDate = pil.pilTransfers
-    .filter(t => t.fromEstablishmentId === id)
-    .reduce((date, t) => {
-      return ((!date || t.createdAt > date) && t.createdAt > pil.startDate) ? t.createdAt : date;
-    }, null) || pil.revocationDate;
-
-  pil.establishmentId = id;
-  return pil;
-};
-
-router.get('/billable',
-  permissions('establishment.licenceFees'),
-  (req, res, next) => {
-    const { limit, offset, sort = {} } = req.query;
-    const { PIL } = req.models;
-
-    const year = parseInt(req.query.year, 10);
-    const start = `${year}-04-06`;
-    const end = `${year + 1}-04-05`;
-
-    const params = {
-      establishmentId: req.establishment.id,
-      start,
-      end
-    };
-    let query = PIL.query().billable(params);
-    query = PIL.orderBy({ query, sort: { ...sort, column: sort.column || 'profile.lastName' } });
-    Promise.resolve()
-      .then(() => PIL.paginate({ query, limit, offset }))
-      .then(pils => {
-        res.meta.count = pils.total;
-        res.meta.total = pils.total;
-        res.meta.startDate = start;
-        res.meta.endDate = end;
-        res.response = pils.results.map(clean(req.establishment.id, start, end));
-      })
-      .then(() => next('router'))
-      .catch(next);
-  }
-);
 
 router.param('pilId', (req, res, next, id) => {
   if (!isUUID(id)) {
